Reestructuracion completada
