+++ frontend/src/App.jsx
@@ -1,25 +1,26 @@
-import React from 'react';
-import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import Navbar from './components/navbar/navbar';
-import HomePage from './pages/homepage'; //Landing page Venta
-import Resgistrarse from './components/resgistrarse';
-import IniciarSesion from './components/iniciarSesion';
-import SobreNosotros from './pages/sobreNosotros'; // ruta quienes somos
-import PoliticasPrivacidad from './components/politicasPrivacidad'; // ruta politica de privacidad
-import TerminosCond from './components/terminosCond'; // ruta terminos y condiciones
+import React from "react";
+import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
+// import Navbar from "./components/navbar/navbar";
+import HomePage from "./pages/Home/homepage"; //Landing page Venta
+// import Resgistrarse from "./components/resgistrarse";
+// import IniciarSesion from "./components/iniciarSesion";
+// import SobreNosotros from "./pages/sobreNosotros"; // ruta quienes somos
+// import PoliticasPrivacidad from "./components/politicasPrivacidad"; // ruta politica de privacidad
+// import TerminosCond from "./components/terminosCond"; // ruta terminos y condiciones
 
 //Importaciones de Luis (Autenticacion)
-import Login from "./pages/Login";
-import Register from "./pages/Register";
-import PasswordRecovery from "./pages/PasswordRecovery";
-import EmailRecovery from "./pages/EmailRecovery";
-import PasswordReset from "./pages/PasswordReset";
-import Terms from "./pages/Terms";
-import PrivacyPolicy from "./pages/PrivacyPolicy";
+import Login from "./pages/Auth/Login";
+import Register from "./pages/Auth/Register";
+import PasswordRecovery from "./pages/Auth/PasswordRecovery";
+import EmailRecovery from "./pages/Auth/EmailRecovery";
+import PasswordReset from "./pages/Auth/PasswordReset";
+import Terms from "./pages/Legal/Terms";
+import PrivacyPolicy from "./pages/Legal/PrivacyPolicy";
 import Simulation from "./pages/Simulation";
+// import Error404 from "./pages/Error404";
+
 
 const App = () => {
-
   return (
     <BrowserRouter>
       <Routes>
@@ -40,7 +41,6 @@ const App = () => {
         <Route path="/simulacion" element={<Simulation />} />
       </Routes>
     </BrowserRouter>
-
   );
 };
