+++ frontend/src/components/common/AnimatedContainer.jsx
