+++ frontend/src/components/common/FormContainer.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from "../assets/logo.png";
+import logo from "../../assets/logo.png";
 
 const FormContainer = ({ children }) => {
   return (
