+++ frontend/src/components/common/PasswordInput.jsx
