+++ frontend/src/components/functionalities/passwordValidation.jsx
