+++ frontend/src/components/home/footer.jsx
@@ -1,0 +1,77 @@
+import { Link, useSubmit } from "react-router-dom";
+import React from "react";
+import logoComplete from "../../../public/img/logoCompleteWhite.svg";
+
+const Footer = () => {
+  return (
+    <div className="relative py-10 bg-black text-white">
+      <div className="container mx-auto px-4">
+        <div className="flex flex-col md:flex-row justify-between items-center gap-8 mb-8">
+          {/* Logo */}
+          <div className="mb-6 md:mb-0">
+            <img 
+              src={logoComplete} 
+              alt="Logo de la empresa" 
+              className="h-25 w-auto "
+            />
+          </div>
+          
+          {/* Redes Sociales */}
+          <div className="flex items-center gap-6">
+            <a href="#" className="hover:text-gray-300 transition-colors duration-300">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="currentColor"
+                viewBox="0 0 320 512"
+                className="h-6 w-6">
+                <path
+                  d="M80 299.3V512H196V299.3h86.5l18-97.8H196V166.9c0-51.7 20.3-71.5 72.7-71.5c16.3 0 29.4 .4 37 1.2V7.9C291.4 4 256.4 0 236.2 0C129.3 0 80 50.5 80 159.4v42.1H14v97.8H80z" />
+              </svg>
+              <span className="sr-only">Facebook</span>
+            </a>
+            <a href="#" className="hover:text-gray-300 transition-colors duration-300">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="currentColor"
+                viewBox="0 0 448 512"
+                className="h-6 w-6">
+                <path
+                  d="M224.1 141c-63.6 0-114.9 51.3-114.9 114.9s51.3 114.9 114.9 114.9S339 319.5 339 255.9 287.7 141 224.1 141zm0 189.6c-41.1 0-74.7-33.5-74.7-74.7s33.5-74.7 74.7-74.7 74.7 33.5 74.7 74.7-33.6 74.7-74.7 74.7zm146.4-194.3c0 14.9-12 26.8-26.8 26.8-14.9 0-26.8-12-26.8-26.8s12-26.8 26.8-26.8 26.8 12 26.8 26.8zm76.1 27.2c-1.7-35.9-9.9-67.7-36.2-93.9-26.2-26.2-58-34.4-93.9-36.2-37-2.1-147.9-2.1-184.9 0-35.8 1.7-67.6 9.9-93.9 36.1s-34.4 58-36.2 93.9c-2.1 37-2.1 147.9 0 184.9 1.7 35.9 9.9 67.7 36.2 93.9s58 34.4 93.9 36.2c37 2.1 147.9 2.1 184.9 0 35.9-1.7 67.7-9.9 93.9-36.2 26.2-26.2 34.4-58 36.2-93.9 2.1-37 2.1-147.8 0-184.8zM398.8 388c-7.8 19.6-22.9 34.7-42.6 42.6-29.5 11.7-99.5 9-132.1 9s-102.7 2.6-132.1-9c-19.6-7.8-34.7-22.9-42.6-42.6-11.7-29.5-9-99.5-9-132.1s-2.6-102.7 9-132.1c7.8-19.6 22.9-34.7 42.6-42.6 29.5-11.7 99.5-9 132.1-9s102.7-2.6 132.1 9c19.6 7.8 34.7 22.9 42.6 42.6 11.7 29.5 9 99.5 9 132.1s2.7 102.7-9 132.1z" />
+              </svg>
+              <span className="sr-only">Instagram</span>
+            </a>
+            <a href="#" className="hover:text-gray-300 transition-colors duration-300">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="currentColor"
+                viewBox="0 0 512 512"
+                className="h-6 w-6">
+                <path
+                  d="M389.2 48h70.6L305.6 224.2 487 464H345L233.7 318.6 106.5 464H35.8L200.7 275.5 26.8 48H172.4L272.9 180.9 389.2 48zM364.4 421.8h39.1L151.1 88h-42L364.4 421.8z" />
+              </svg>
+              <span className="sr-only">Twitter</span>
+            </a>
+          </div>
+        </div>
+        
+        {/* Parte inferior con derechos y enlaces */}
+        <div className="flex flex-col md:flex-row justify-between items-center text-sm text-gray-400">
+          <div className="mb-4 md:mb-0">
+            &copy; {new Date().getFullYear()} ATLAS SOFT. Todos los derechos reservados.
+          </div>
+          
+          <div className="flex flex-wrap justify-center gap-4 md:gap-6">
+            <Link to="/politica-de-privacidad" state={{ from: "/" }} className="hover:text-white transition-colors duration-300">
+              Política de Privacidad
+            </Link>
+            <Link to="/terminos-y-condiciones" state={{ from: "/" }} className="hover:text-white transition-colors duration-300">
+              Términos y Condiciones
+            </Link>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Footer;
