+++ frontend/src/pages/Simulation.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "../components/Button";
+import Button from "../components/common/Button";
 import { Link } from "react-router-dom";
 
 const Simulation = () => {
